Rename filename prefix from logger to log
