+++ srcStatic/Log/LogDebug.cpp
@@ -1,4 +1,4 @@
-#include "LoggerDebug.h"
+#include "LogDebug.h"
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
