+++ srcStatic/Log/LogDebug.h
