+++ srcStatic/Log/LogDistributor.cpp
@@ -1,4 +1,4 @@
-#include "LoggerDistributor.h"
+#include "LogDistributor.h"
 #include <utility>
