+++ srcStatic/Log/LogMessageBox.cpp
@@ -1,4 +1,4 @@
-#include "LoggerMessageBox.h"
+#include "LogMessageBox.h"
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
