+++ srcStatic/Log/LogMessageBox.h
