+++ srcStatic/op2extStatic.vcxproj
@@ -140,10 +140,10 @@
     <ClCompile Include="GameModules\IpDropDown.cpp" />
     <ClCompile Include="IniFile.cpp" />
     <ClCompile Include="Log.cpp" />
-    <ClCompile Include="Log/LoggerDebug.cpp" />
-    <ClCompile Include="Log/LoggerDistributor.cpp" />
-    <ClCompile Include="Log/LoggerFile.cpp" />
-    <ClCompile Include="Log/LoggerMessageBox.cpp" />
+    <ClCompile Include="Log\LogDebug.cpp" />
+    <ClCompile Include="Log\LogDistributor.cpp" />
+    <ClCompile Include="Log\LogFile.cpp" />
+    <ClCompile Include="Log\LogMessageBox.cpp" />
     <ClCompile Include="OP2Memory.cpp" />
     <ClCompile Include="ResourceSearchPath.cpp" />
     <ClCompile Include="StringConversion.cpp" />
@@ -168,10 +168,10 @@
     <ClInclude Include="LocalResource.h" />
     <ClInclude Include="Log.h" />
     <ClInclude Include="Log\LogDestination.h" />
-    <ClInclude Include="Log/LoggerDebug.h" />
-    <ClInclude Include="Log/LoggerDistributor.h" />
-    <ClInclude Include="Log/LoggerFile.h" />
-    <ClInclude Include="Log/LoggerMessageBox.h" />
+    <ClInclude Include="Log\LogDebug.h" />
+    <ClInclude Include="Log\LogDistributor.h" />
+    <ClInclude Include="Log\LogFile.h" />
+    <ClInclude Include="Log\LogMessageBox.h" />
     <ClInclude Include="OP2Memory.h" />
     <ClInclude Include="ResourceSearchPath.h" />
     <ClInclude Include="StringConversion.h" />
